### 2024-07-17 #7000 @editorialbot
[PRE REVIEW]: HyperCoast: A Python Package for Visualizing and Analyzing Hyperspectral Data in Coastal Environments
**Submitting author:** @giswqs (<a href="http://orcid.org/0000-0001-5437-4073">Qiusheng Wu</a>)
**Repository:** https://github.com/opengeos/HyperCoast
**Branch with paper.md** (empty if default branch):
**Version:** v0.7.0
