### 2024-07-17 #7000 @kthyng
- [ ] Clear research application/in scope
