### 2024-07-19 b75ca58
Add support for generic dataset
