### 2024-07-19 c4dc42d
Fix add_raster bug
