+++ hypercoast/hypercoast.py
@@ -160,9 +160,12 @@ class Map(leafmap.Map):
             **kwargs,
         )
 
-        da = rxr.open_rasterio(source, **open_args)
-        dims = da.dims
-        da = da.transpose(dims[1], dims[2], dims[0])
+        if isinstance(source, str):
+            da = rxr.open_rasterio(source, **open_args)
+            dims = da.dims
+            da = da.transpose(dims[1], dims[2], dims[0])
+        elif isinstance(source, xr.DataArray):
+            da = source
         xds = da.to_dataset(name="data")
         self.cog_layer_dict[layer_name]["xds"] = xds
         self.cog_layer_dict[layer_name]["hyper"] = "COG"
