### 2024-07-19 4230a57
Fix add_raster bug
