### 2024-07-19 25fa00a
Add support for generic dataset (#78)
* Add support for generic dataset

* Add spectral profile

* Fix add_raster bug

* Fix add_raster bug

* Change COG to XARRAY
