### 2024-07-20 dbfccf9
Clarify license info
