+++ hypercoast/aviris.py
@@ -1,7 +1,8 @@
 """This module contains functions to read and process NASA AVIRIS hyperspectral data.
 More info about the data can be found at https://aviris.jpl.nasa.gov.
-The source code is adapted from https://bit.ly/4bRCgqs. Credit goes to the
-original authors.
+A portion of the source code is adapted from the jjmcnelis/aviris-ng-notebooks repository
+available at https://bit.ly/4bRCgqs. It is licensed under the MIT License. Credit goes to the
+original author Jack McNelis.
 """
 
 import rioxarray
