### 2024-07-20 5e1002f
Add multispectral notebook example
