+++ mkdocs.yml
@@ -102,6 +102,7 @@ nav:
           - examples/temperature.ipynb
           - examples/pace_oci_l1.ipynb
           - examples/pace_oci_l2.ipynb
+          - examples/multispectral.ipynb
     - Workshops:
           - workshops/emit.ipynb
     - API Reference:
