+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "HyperCoast"
-version = "0.7.0"
+version = "0.7.1"
 dynamic = [
     "dependencies",
 ]
@@ -54,7 +54,7 @@ universal = true
 
 
 [tool.bumpversion]
-current_version = "0.7.0"
+current_version = "0.7.1"
 commit = true
 tag = true
 
