### 2024-07-23 826954a
Add run_acolite function
