### 2024-07-23 #7000 @platipodium
> 👋🏼 @platipodium and @malmans2, Would you like to review this submission to the Journal for Open Source Software?

Hi @cheginit, I can take on this review!
