### 2024-07-23 5ba04a3
Add acolite notebook example
