+++ hypercoast/common.py
@@ -766,15 +766,16 @@ def extract_spectral(
 
 def download_acolite(outdir: str = ".", platform: Optional[str] = None) -> str:
     """
-    Downloads the ACOLITE release based on the OS platform and extracts it to the specified directory.
+    Downloads the Acolite release based on the OS platform and extracts it to the specified directory.
+    For more information, see the Acolite manual https://github.com/acolite/acolite/releases.
 
     Args:
-        outdir (str): The output directory where the file will be downloaded and extracted.
-        platform (Optional[str]): The platform for which to download ACOLITE. If None, the current system platform is used.
+        outdir (str): The output directory where the file will be Acolite and extracted.
+        platform (Optional[str]): The platform for which to download acolite. If None, the current system platform is used.
                                   Valid values are 'linux', 'darwin', and 'windows'.
 
     Returns:
-        str: The path to the extracted ACOLITE directory.
+        str: The path to the extracted Acolite directory.
 
     Raises:
         Exception: If the platform is unsupported or the download fails.
@@ -845,7 +846,7 @@ def run_acolite(
     acolite_dir: str,
     settings_file: Optional[str] = None,
     input_file: Optional[str] = None,
-    output: Optional[str] = None,
+    out_dir: Optional[str] = None,
     polygon: Optional[str] = None,
     l2w_parameters: Optional[str] = None,
     rgb_rhot: bool = True,
@@ -855,39 +856,53 @@ def run_acolite(
     **kwargs: Any,
 ) -> None:
     """
-    Runs the ACOLITE software for atmospheric correction and water quality retrieval.
+    Runs the Acolite software for atmospheric correction and water quality retrieval.
+    For more information, see the Acolite manual https://github.com/acolite/acolite/releases
 
-    This function constructs and executes a command to run the ACOLITE software with the specified
-    parameters. It supports running ACOLITE with a settings file or with individual parameters
+    This function constructs and executes a command to run the Acolite software with the specified
+    parameters. It supports running Acolite with a settings file or with individual parameters
     specified directly. Additional parameters can be passed as keyword arguments.
 
     Args:
-        acolite_dir (str): The directory where ACOLITE is installed.
-        settings_file (Optional[str], optional): The path to the ACOLITE settings file. If provided,
+        acolite_dir (str): The directory where Acolite is installed.
+        settings_file (Optional[str], optional): The path to the Acolite settings file. If provided,
             other parameters except `verbose` are ignored. Defaults to None.
         input_file (Optional[str], optional): The path to the input file for processing. Defaults to None.
-        output (Optional[str], optional): The directory where output files will be saved. Defaults to None.
+        out_dir (Optional[str], optional): The directory where output files will be saved. Defaults to None.
         polygon (Optional[str], optional): The path to a polygon file for spatial subset. Defaults to None.
         l2w_parameters (Optional[str], optional): Parameters for L2W processing. Defaults to None.
         rgb_rhot (bool, optional): Flag to generate RGB images using rhot. Defaults to True.
         rgb_rhos (bool, optional): Flag to generate RGB images using rhos. Defaults to True.
         map_l2w (bool, optional): Flag to map L2W products. Defaults to True.
         verbose (bool, optional): If True, prints the command output; otherwise, suppresses it. Defaults to True.
-        **kwargs (Any): Additional command line arguments to pass to ACOLITE.
[…]
