### 2024-07-23 #7025 @editorialbot
[REVIEW]: HyperCoast: A Python Package for Visualizing and Analyzing Hyperspectral Data in Coastal Environments
**Submitting author:** @giswqs (<a href="http://orcid.org/0000-0001-5437-4073">Qiusheng Wu</a>)
**Repository:** https://github.com/opengeos/HyperCoast
**Branch with paper.md** (empty if default branch):
**Version:** v0.7.8
