### 2024-07-23 612aba1
Add support for acolite (#81)
* Add support for acolite

* Add run_acolite function

* Fix typos

* Add acolite notebook example

* Update README
