### 2024-07-23 99180f7
Add support for acolite (#81)
* Add support for acolite

* Add run_acolite function

* Fix typos

* Add acolite notebook example

* Update README
