### 2024-07-23 opengeos/HyperCoast#82 @platipodium
Use license management framework
### Description

As you claim to use work under different licenses within your source code, a license management framework might be helpful.

* Please consider to use https://reuse.software or similar to do so.

It should be easy to implement and can also be added to your .pre-commit or CI pipeline.

This is part of the review for the JOSS paper at https://github.com/openjournals/joss-reviews/issues/7025
