### 2024-07-23 opengeos/HyperCoast#82 @platipodium
This is a mere suggestion.  If you have good reason not to follow this route, please explain why not.
