### 2024-07-23 opengeos/HyperCoast#82 @giswqs
Good suggestion. I will implement it. Thanks.
