### 2024-07-23 96a37f9
Add resue license management framework
