+++ .pre-commit-config.yaml
@@ -17,13 +17,22 @@ repos:
           - id: black-jupyter
             language_version: python3
 
-    # - repo: https://github.com/codespell-project/codespell
-    #   rev: v2.2.6
-    #   hooks:
-    #       - id: codespell
-    #         args: [--toml, pyproject-codespell.precommit-toml]
+    - repo: https://github.com/codespell-project/codespell
+      rev: v2.3.0
+      hooks:
+          - id: codespell
+            args:
+                [
+                    "--ignore-words-list=gis,timeseries,sav",
+                    "--skip=*.json,*.csv",
+                ]
 
     - repo: https://github.com/kynan/nbstripout
       rev: 0.7.1
       hooks:
           - id: nbstripout
+
+    - repo: https://github.com/fsfe/reuse-tool
+      rev: v4.0.3
+      hooks:
+          - id: reuse
