### 2024-07-23 ee0691c
Add resue license management framework (#83)
* Add resue license management framework

* Update REUSE.toml
