### 2024-07-23 opengeos/HyperCoast#82 @giswqs
This has been implemented in #83, adding a `REUSE.toml` and pre-commit. Let me know if additional changes are needed. Thanks.
