### 2024-07-23 opengeos/HyperCoast#82 @platipodium
Please reopen, as the test does not pass.

```
reuse lint
# UNUSED LICENSES

The following licenses are not used:
* Apache-2.0
* MIT

# MISSING COPYRIGHT AND LICENSING INFORMATION

The following files have no copyright and licensing information:
* .editorconfig
* .github/ISSUE_TEMPLATE/bug_report.md
* .github/ISSUE_TEMPLATE/config.yml
* .github/ISSUE_TEMPLATE/feature_request.md
* .github/dependabot.yml
* .github/workflows/docs-build.yml
* .github/workflows/docs.yml
* .github/workflows/draft-pdf.yml
* .github/workflows/installation.yml
* .github/workflows/macos.yml
* .github/workflows/pypi.yml
* .github/workflows/ubuntu.yml
* .github/workflows/windows.yml
* .gitignore
* .pre-commit-config.yaml
* MANIFEST.in
* README.md
* REUSE.toml
* docs/CNAME
* docs/aviris.md
* docs/changelog.md
* docs/changelog_update.py
* docs/common.md
* docs/contributing.md
* docs/desis.md
* docs/emit.md
* docs/examples/acolite.ipynb
* docs/examples/aviris.ipynb
* docs/examples/chlorophyll_a.ipynb
* docs/examples/desis.ipynb
* docs/examples/ecostress.ipynb
* docs/examples/emit.ipynb
* docs/examples/image_cube.ipynb
* docs/examples/image_slicing.ipynb
* docs/examples/multispectral.ipynb
* docs/examples/neon.ipynb
* docs/examples/overview.md
* docs/examples/pace.ipynb
* docs/examples/pace_oci_l1.ipynb
* docs/examples/pace_oci_l2.ipynb
* docs/examples/search_data.ipynb
* docs/examples/temperature.ipynb
* docs/hypercoast.md
* docs/index.md
* docs/installation.md
* docs/neon.md
* docs/overrides/main.html
* docs/pace.md
* docs/ui.md
* docs/usage.md
* docs/workshops/emit.ipynb
* hypercoast/__init__.py
* hypercoast/aviris.py
* hypercoast/common.py
* hypercoast/desis.py
* hypercoast/emit.py
* hypercoast/hypercoast.py
* hypercoast/neon.py
* hypercoast/pace.py
* hypercoast/ui.py
* paper/paper.bib
* paper/paper.md
* pyproject.toml
* requirements.txt
* requirements_dev.txt
* tests/__init__.py
* tests/test_aviris.py
* tests/test_emit.py
* tests/test_hypercoast.py
* tests/test_neon.py
* tests/test_pace.py

The following files have no licensing information:
* mkdocs.yml

# SUMMARY

* Bad licenses: 0
* Deprecated licenses: 0
* Licenses without file extension: 0
* Missing licenses: 0
* Unused licenses: Apache-2.0, MIT
* Used licenses: 0
* Read errors: 0
* files with copyright information: 1 / 72
* files with license information: 0 / 72

Unfortunately, your project is not compliant with version 3.0 of the REUSE Specification :-(

# RECOMMENDATIONS

* Fix unused licenses: At least one of the license text files in 'LICENSES' is
  not referenced by any file, e.g. by an 'SPDX-License-Identifier' tag. Please
  make sure that you either tag the accordingly licensed files properly, or
  delete the unused license text if you are sure that no file or code snippet is
  licensed as such.
* Fix missing copyright/licensing information: For one or more files, the tool
  cannot find copyright and/or licensing information. You typically do this by
  adding 'SPDX-FileCopyrightText' and 'SPDX-License-Identifer' tags to each
  file. The tutorial explains additional ways to do this:
  <https://reuse.software/tutorial/>
```
