### 2024-07-23 opengeos/HyperCoast#82 @platipodium
`REUSE.toml` does not do it :=).   You can use the `.reuse/dep5` file for summarily assigning copyright and license to multiple files in a directory.  It uses similar (but not identical) info as your existing `REUSE.toml`
