### 2024-07-23 opengeos/HyperCoast#82 @platipodium
Sorry, my reuse was outdated!  (version 3.0 spec).  Of course it is acceptable with the most recent change.   I was not aware of the deprecation of .dep5 in v3.2, so thanks for the heads up!  Still try to add the information to file headers when you happen to edit them (there is also a bulk addition option in the reuse command line).

* please add reuse>=3.2 to requirements_dev.txt
