### 2024-07-23 opengeos/HyperCoast#82 @platipodium
... actually. reuse 3.0.2. does not support spec 3.2 yet, so people like me with installed reuse (conda has 3.0.2) will encounter this problem.  Need to handle this somehow through requirments_dev.txt.  How did you do it?
