### 2024-07-23 opengeos/HyperCoast#82 @platipodium
According to https://repology.org/project/reuse/versions, many existing package manager still only support reuse 3.0.2, so not compatible with spec 3.2 (supported by reuse tool 4.x).  But I guess it's okay for your project.
