### 2024-07-23 opengeos/HyperCoast#82 @giswqs
I have added file headers and the reuse>3.2 requirement in https://github.com/opengeos/HyperCoast/pull/84.

The requirments_dev.txt is for pip install only. I just used `pip install reuse` in a conda env. It works fine.
