+++ hypercoast/ui.py
@@ -1,3 +1,7 @@
+# SPDX-FileCopyrightText: 2024 Qiusheng Wu <giswqs@gmail.com>
+#
+# SPDX-License-Identifier: MIT
+
 """This module contains the user interface for the hypercoast package.
 """
 
