### 2024-07-23 opengeos/HyperCoast#82 @platipodium
Just to confirm from my side:

```
19:08:52 ❯ LANG=EN_US reuse lint

# SUMMARY

* Bad licenses: 0
* Deprecated licenses: 0
* Licenses without file extension: 0
* Missing licenses: 0
* Unused licenses: 0
* Used licenses: Apache-2.0, MIT
* Read errors: 0
* Files with copyright information: 73 / 73
* Files with license information: 73 / 73

Congratulations! Your project is compliant with version 3.2 of the REUSE Specification :-)
```
