### 2024-07-23 opengeos/HyperCoast#82 @giswqs
Fix in https://github.com/opengeos/HyperCoast/pull/85. Thanks.
