### 2024-07-23 opengeos/HyperCoast#87 @platipodium
Try to reduce literature
### Description

The paper has 2 pages of literature for 1 page or text content.  Though most of the citations are in principle fine and relevant, some could possibly be removed.

Candidate from my perspective:
 * remove sentence "It enables researchers and environmental managers to gain deeper insights into the dynamic processes occurring in aquatic environments (Liu et al., 2021, 2023)." and refs to Liu

but there may be more...

Linked to https://github.com/openjournals/joss-reviews/issues/7025
