### 2024-07-23 opengeos/HyperCoast#88 @platipodium
Explain hyperspectral and (possibly) other jargon in paper
### Description

Assume that your paper should be readable by a person not necessarily experienced with spectral analysis.  You state yourself that your target audience could be "managers".  Therefore, avoid or explain jargon terms.  Especially, add (e.g. in the summary at first occurrence of the word hyper spectral) an explanation;  what is spectral, what is hyperspectral, and what is it good for!

There may be more jargon, so please go over your paper with that "lay" perspective.

----

Linked to https://github.com/openjournals/joss-reviews/issues/7025
