### 2024-07-23 #7025 @platipodium
@cheginit what do you think of the 2 pages references versus 1-2 pages of text?  Is it ok or should the authors try to balance more and reduce refs?
