### 2024-07-23 opengeos/HyperCoast#87 @platipodium
As in the main review thread.  D. Katz said that JOSS does not limit refs.  So you can disregard this comment.  (But I still think the one sentence quoted above is not relevant)
