### 2024-07-23 28135fc
Decapitalize Pringle et al title (#89)
