+++ paper/paper.bib
@@ -395,8 +395,7 @@
 }
 
 @ARTICLE{Pringle2001,
-  title     = "{HYDROLOGIC CONNECTIVITY AND THE MANAGEMENT OF BIOLOGICAL
-               RESERVES: A GLOBAL PERSPECTIVE}",
+  title     = "{Hydrologic connectivity and the management of biological reserves: a global perspective}",
   author    = "Pringle, Catherine M",
   journal   = "Ecological applications: a publication of the Ecological Society
                of America",
