### 2024-07-23 f43444b
Remove sentence and citations
