+++ paper/paper.bib
@@ -26,41 +26,6 @@
   doi       = "10.21105/joss.01450"
 }
 
-@ARTICLE{Liu2021,
-  title    = "{Biogeographical trends in phytoplankton community size structure
-              using adaptive sentinel 3-OLCI chlorophyll a and spectral
-              empirical orthogonal functions in the estuarine-shelf waters of
-              the northern Gulf of Mexico}",
-  author   = "Liu, B and D'Sa, Eurico J and Maiti, Kanchan and
-              Rivera-Monroy, Victor H and Xue, Zuo",
-  journal  = "Remote sensing of environment",
-  volume   =  252,
-  pages    = "112154",
-  year     =  2021,
-  url      = "https://www.sciencedirect.com/science/article/pii/S0034425720305277",
-  keywords = "Phytoplankton size fraction; Chlorophyll a; Adaptive; EOF;
-              Sentinel 3-OLCI; Northern Gulf of Mexico",
-  issn     = "0034-4257",
-  doi      = "10.1016/j.rse.2020.112154"
-}
-
-@ARTICLE{Liu2023,
-  title     = "{Quantifying the Potential Contribution of Submerged Aquatic
-               Vegetation to Coastal Carbon Capture in a Delta System from
-               Field and Landsat 8/9-Operational …}",
-  author    = "Liu, B and Sevick, T and Jung, H and Kiskaddon, E and
-               Carruthers, T",
-  abstract  = "Submerged aquatic vegetation (SAV) are highly efficient at
-               carbon sequestration and, despite their relatively small
-               distribution globally, are recognized as a potentially valuable
-               …",
-  journal   = "Remote Sensing",
-  publisher = "mdpi.com",
-  year      =  2023,
-  url       = "https://www.mdpi.com/2072-4292/15/15/3765",
-  doi       = "10.3390/rs15153765"
-}
-
 @MISC{De_La_Pena2017,
   title     = "{hyperspy/hyperspy: HyperSpy 1.3}",
   author    = "De La Pe{\~n}a, Francisco and Ostasevicius, Tomas and Tonaas
