### 2024-07-23 772b1d2
Remove sentence and citations (#90)
