### 2024-07-23 9a7397b
Explain hyperspectral and other jargon
