### 2024-07-31 d4c2556
Add xlim parameter for the spectral widget (#93)
* Add xlim parameter for the spectral widget

* Make download quiet
