+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "HyperCoast"
-version = "0.7.2"
+version = "0.7.3"
 dynamic = [
     "dependencies",
 ]
@@ -54,7 +54,7 @@ universal = true
 
 
 [tool.bumpversion]
-current_version = "0.7.2"
+current_version = "0.7.3"
 commit = true
 tag = true
 
