### 2024-08-02 a2292e7
Remove AVIRIS image bounds
