+++ hypercoast/hypercoast.py
@@ -664,7 +664,6 @@ class Map(leafmap.Map):
             **kwargs,
         )
 
-        xds.attrs["bounds"] = self.cog_layer_dict[layer_name]["bounds"]
         self.cog_layer_dict[layer_name]["xds"] = xds
         self.cog_layer_dict[layer_name]["hyper"] = "AVIRIS"
         self._update_band_names(layer_name, wavelengths)
