+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "HyperCoast"
-version = "0.7.3"
+version = "0.7.4"
 dynamic = [
     "dependencies",
 ]
@@ -54,7 +54,7 @@ universal = true
 
 
 [tool.bumpversion]
-current_version = "0.7.3"
+current_version = "0.7.4"
 commit = true
 tag = true
 
