### 2024-08-03 798d8b6
ci(Mergify): configuration update (#99)
