+++ requirements.txt
@@ -9,5 +9,6 @@ netcdf4
 numpy<2.0.0
 rioxarray
 s3fs
+scikit-learn
 scipy
 xarray
