+++ mkdocs.yml
@@ -105,6 +105,7 @@ nav:
           - examples/pace_oci_l1.ipynb
           - examples/pace_oci_l2.ipynb
           - examples/multispectral.ipynb
+          - examples/pca.ipynb
     - Workshops:
           - workshops/emit.ipynb
     - API Reference:
