### 2024-08-03 f3a7863
Update mergify (#101)
* Update mergify

* Fix typo
