+++ .github/workflows/docs-build.yml
@@ -6,7 +6,7 @@ on:
             - master
 
 jobs:
-    deploy:
+    test-docs-build:
         runs-on: ubuntu-latest
         steps:
             - uses: actions/checkout@v4
