### 2024-08-03 698e440
Update mergify (#101)
* Update mergify

* Fix typo
