+++ .mergify.yml
@@ -1,37 +1,40 @@
 pull_request_rules:
-  - name: CI automerge when passing all test
-    conditions:
-      # If the requirement are satisfied
-      # The bot will trigger automerge
-      - label=ready-to-merge
-      - check-success=ubuntu-latest (3.9)
-      - check-success=ubuntu-latest (3.10)
-      - check-success=ubuntu-latest (3.11)
-      - check-success=macOS-latest (3.11)
-      # Approval by reviewer at least one maintainer
-      # - "#approved-reviews-by>=1"
-      - base=main
-    actions:
-      comment:
-        # message if the pull request success to merge
-        message: Thanks for the contribution @{{author}} 🤩
-      label:
-        # adding label to merged pull request
-        add:
-          - already reviewed
-      merge:
-        method: squash
+    - name: CI automerge when passing all test
+      conditions:
+          # If the requirement are satisfied
+          # The bot will trigger automerge
+          - label=ready-to-merge
+          - check-success=ubuntu-latest (3.9)
+          - check-success=ubuntu-latest (3.10)
+          - check-success=ubuntu-latest (3.11)
+          - check-success=macOS-latest (3.11)
+          - check-success=test-windows
+          - check-success=test-docs-build
+          - check-success=test-linux-installation
+          # Approval by reviewer at least one maintainer
+          # - "#approved-reviews-by>=1"
+          - base=main
+      actions:
+          comment:
+              # message if the pull request success to merge
+              message: Thanks for the contribution @{{author}} 🤩
+          label:
+              # adding label to merged pull request
+              add:
+                  - already reviewed
+          merge:
+              method: squash
 
-  - name: information about conflicts pull request
-    conditions:
-      # If the requirement are satisfied
-      # The bot will trigger conflict process
-      - conflict
-    actions:
-      comment:
-        # Message if there's conflict on pull request
-        message: your pull request are conflict @{{author}}, please fix it !
-      label:
-        # adding label on pull request if get trigger
-        add:
-          - conflicts pull request
+    - name: information about conflicts pull request
+      conditions:
+          # If the requirement are satisfied
+          # The bot will trigger conflict process
+          - conflict
+      actions:
+          comment:
+              # Message if there's conflict on pull request
+              message: Your pull request are conflict @{{author}}, please fix it!
+          label:
+              # adding label on pull request if get trigger
+              add:
+                  - conflicts pull request
