### 2024-08-06 f2969e9
Clean up file path in notebook
