### 2024-08-08 opengeos/HyperCoast#110 @platipodium
Add information on legal requirements need for contributions
### Description

As part of the review for JOSS at https://github.com/openjournals/joss-reviews/issues/7025 we are asked whether the community information is complete.

The project gives useful and elaborate information on How to Contribute in its `contributing.md` file as part of the documentation and source.

### Missing information

The legal aspects of a third-party contributions are missing.  Do you need copyright transfer?  Do you have a contributor license agreement?  This is as of yet a much neglected (but legally required) aspect of Open Source collaborative development, please add statements to guide your potential contributors to clarify the copyright of their contributions to your project.

You may have a look at the Harmony Agreements for more explanations and potential solutions. See https://www.harmonyagreements.org
