### 2024-08-08 #7025 @cheginit
@platipodium Thanks for paying attention to this. These labels are automatically generated based on the "Software report", since there are Tex files in the repo, the label is added.
