### 2024-08-08 opengeos/HyperCoast#110 @giswqs
Thank you for the suggestion. Do you have recommendations for sample repositories that use Harmony Agreements?
