### 2024-08-08 opengeos/HyperCoast#110 @giswqs
What do you think about this Contributor License Agreement (CLA) assistant?
https://github.com/cla-assistant/cla-assistant
