### 2024-08-09 opengeos/HyperCoast#110 @platipodium
https://contributoragreements.org/ seems also useful, though I have not used it yet for production.  It seems to be very much in line with what harmony agreements suggest.  What it adds is the link to institutional license policies (if they are publicly accessible)
