### 2024-08-09 opengeos/HyperCoast#110 @platipodium
> Do you have recommendations for sample repositories that use Harmony Agreements?

Actually I don't.  In our latest repo we have the individual CLA built from the Harmony Agreements https://codebase.helmholtz.cloud/mussel/netlogo-northsea-species/-/blob/main/doc/contributing/contributing-license.md, but we have not yet had to apply it, as the contribution legal issues with the current contributor team in this (and other research projects I run) are dealt with  by separate inter-institutional memoranda.
