+++ .github/workflows/cla.yml
@@ -26,7 +26,7 @@ jobs:
                   PERSONAL_ACCESS_TOKEN: ${{ secrets.PERSONAL_ACCESS_TOKEN }}
               with:
                   path-to-signatures: "docs/signatures/cla.json"
-                  path-to-document: "https://github.com/cla-assistant/github-action/blob/master/docs/SAPCLA.md" # e.g. a CLA or a DCO document
+                  path-to-document: "https://github.com/cla-assistant/github-action/blob/master/docs/cla.md" # e.g. a CLA or a DCO document
                   # branch should not be protected
                   branch: "main"
                   allowlist: user1,bot*
