### 2024-08-09 opengeos/HyperCoast#110 @giswqs
@platipodium Can you take a look at #111 to see if it address your comments?
