### 2024-08-10 opengeos/HyperCoast#110 @platipodium
OK, great, that certainly addresses the individual CLA.

* How would you deal with entity CLAs?
* Have you tested the CLA GitHub workflow?
