### 2024-08-10 opengeos/HyperCoast#110 @platipodium
Also link the CLA in your contributing.md
