### 2024-08-11 opengeos/HyperCoast#110 @giswqs
The entity CLA has been added and the links to the CLAs have been added to contributing.md.

The CLA GitHub workflow will be tested once this PR has been merged. Will fix it if it is not working properly.
