### 2024-08-11 0aa5b5c
Add Contributor License Agreement (CLA) assistant (#111)
* Add Contributor License Agreement (CLA) assistant

* Add cla.md

* Update cla.md file path

* Add Entity Contributor License Agreement

* Add Contributor Agreements to contributing.md
