### 2024-08-11 5fbcd1a
Update changelog (#112)
* Update changelog

* Add cla.json

* Remove cla.yml
