### 2024-08-11 opengeos/HyperCoast#110 @giswqs
I installed CLA assistant to the repo through https://cla-assistant.io. You can see from my pull request that the CLA assistant is working properly now.
https://github.com/opengeos/HyperCoast/pull/112#issuecomment-2282898038
