### 2024-08-20 #7025 @alexgleith
Some very minor comments on code quality:

- There are some unused imports, like `numpy` in `desis.py` and others that are probably imported for the context function (not the right word), but should be annotated, like `import hvplot.xarray  # noqa F401`. These are little, but my linter highlights them, and it's cleaner to fix it.
- In the main `hypercoast.py` file, there are `*` imports, which are not recommended `from .aviris import *`
- `spectral` is used, but isn't in the dependencies list
- I'll make a PR with some example code cleanups.

Some opinions from me:

- I don't like that EMIT data has been stored in GitHub, like this: [https://github.com/opengeos/datasets/releases/tag/netcdf](https://github.com/opengeos/datasets/releases/tag/netcdf). I would much prefer that the notebooks point to the data direct at NASA.
- Following on from the first point, wrapping up the EarthData access into the `download_nasa_data` function, which calls `leafmap.nasa_data_download`, which then calls `earthaccess.download` is two abstractions too far! It does make it simple, though. I have an [example over here](https://github.com/auspatious/hyperspectral-notebooks/blob/main/00_EMIT_STAC_Search.ipynb) that uses NASA's STAC API to find EMIT granules and load them, and it's still pretty complex, but what is good is that it uses the open standard STAC to discover granules.
- It's great that the function `emit_xarray` has been adapted to make wavelength a dimension!
- I can't see how this method, in the `emit.ipynb` notebook is working, or if it works. I tried changing the numbers and nothing changed: `m.add("spectral", xlim=(400, 1400))`
- The function `m.spectral_to_df()` structures the dataframe in an unexpected way. I think exporting something with a geometry, or with lon/lat columns and 285 more spectral columns would make more sense, even if the structure is a bit clumsy.

On the review points, I'd like a couple of changes please:

1. For installation, please make sure all dependencies are listed in the pyproject.toml file (`spectral` was not, and neither was `pyvista`)
2. For community guidelines, please include a [contributing file](https://github.com/opendatacube/datacube-core/blob/develop/CONTRIBUTING.md)
