### 2024-08-20 ae3d3b6
Code formatting cleanup
