### 2024-08-20 #7025 @giswqs
@alexgleith Thank you very much for the feedback. I will be addressing your comments in another PR.

> For installation, please make sure all dependencies are listed in the pyproject.toml file (spectral was not, and neither was pyvista)

Regarding the dependencies, `pyvista` has already been listed as an optional dependency in `pyproject.toml`.
`spectral` is not a dependency of the package. We did not use it anywhere in the source code.
https://github.com/opengeos/HyperCoast/blob/07888c48f9edc207cb4ec855986f52bb56f408fb/pyproject.toml#L39
![image](https://github.com/user-attachments/assets/57d4094d-0524-4489-b4a5-29919c5a6232)
