### 2024-08-20 #7025 @giswqs
My mistakes. The `spectral` package is used by the `write_envi` function inherited from the [nasa/EMIT-Data-Resources](https://github.com/nasa/EMIT-Data-Resources) repo. However, the function is not used in our source code. I will remove it from the source code.
