### 2024-08-20 b20fc8a
Code formatting cleanup (#115)
* Code formatting cleanup

* Revert issue with  vs
