### 2024-08-20 #7025 @alexgleith
>pyvista has already been listed as an optional dependency

Ah, sorry, I missed that.
