### 2024-08-20 opengeos/HyperCoast#117 @giswqs
> I don't like that EMIT data has been stored in GitHub, like this: https://github.com/opengeos/datasets/releases/tag/netcdf. I would much prefer that the notebooks point to the data direct at NASA.

The purpose of those three sample files was to make the GitHub Actions unit tests run faster. It takes ~30 seconds to download the sample data, a bit faster than downloading using earthaccess. As we have many notebook examples that use EMIT and PACE data, I try to avoid using my personal NASA EarthData account for authentication to repeatedly downloading these datasets for every commits in a PR.

Additionally, the sample data lowers the barriers for users to try out the HyperCoast features. Users can launch the notebook on Google Colab and explore the functionality without having to sign up for a NASA EarthData account and authenticate it in the notebook. We do provide a separate notebook example ([search_data.ipynb](https://hypercoast.org/examples/search_data/))  that demonstrates how to download NASA EarthData interactively and programmatically.

@alexgleith   Would these arguments be acceptable to you?

![image](https://github.com/user-attachments/assets/50c0e211-ff51-4b0c-bb6e-f54a6ebe1e79)
