### 2024-08-20 31d5394
Remove unused write_envi function and  spectral dependency (#116)
