+++ hypercoast/emit.py
@@ -605,192 +605,6 @@ def band_mask(filepath: str) -> np.ndarray:
     return unpacked_bmask
 
 
-def write_envi(
-    xr_ds: xr.Dataset,
-    output_dir: str,
-    overwrite: Optional[bool] = False,
-    extension: str = ".img",
-    interleave: Optional[str] = "BIL",
-    glt_file: Optional[bool] = False,
-) -> tuple:
-    """
-    Takes an EMIT dataset read into an xarray dataset using the emit_xarray function and writes an ENVI file and header.
-
-    Args:
-        xr_ds (xarray.Dataset): An EMIT dataset read into xarray using the emit_xarray function.
-        output_dir (str): Output directory.
-        overwrite (bool, optional): Overwrite existing file if True. Defaults to False.
-        extension (str, optional): The file extension for the envi formatted file, .img by default. Defaults to ".img".
-        interleave (str, optional): The interleave option for the ENVI file. Defaults to "BIL".
-        glt_file (bool, optional): Also create a GLT ENVI file for later use to reproject. Defaults to False.
-
-    Returns:
-        tuple: A tuple containing:
-            - envi_ds (spectral.io.envi.Image): ENVI file in the output directory.
-            - glt_ds (spectral.io.envi.Image): GLT file in the output directory.
-
-    Raises:
-        Exception: If the data is already orthorectified but a GLT file is still requested.
-    """
-    from spectral.io import envi
-
-    # Check if xr_ds has been orthorectified, raise exception if it has been but GLT is still requested
-    if (
-        "Orthorectified" in xr_ds.attrs.keys()
-        and xr_ds.attrs["Orthorectified"] == "True"
-        and glt_file is True
-    ):
-        raise Exception("Data is already orthorectified.")
-
-    # Typemap dictionary for ENVI files
-    envi_typemap = {
-        "uint8": 1,
-        "int16": 2,
-        "int32": 3,
-        "float32": 4,
-        "float64": 5,
-        "complex64": 6,
-        "complex128": 9,
-        "uint16": 12,
-        "uint32": 13,
-        "int64": 14,
-        "uint64": 15,
-    }
-
-    # Get CRS/geotransform for creation of Orthorectified ENVI file or optional GLT file
-    gt = xr_ds.attrs["geotransform"]
-    mapinfo = (
-        "{Geographic Lat/Lon, 1, 1, "
-        + str(gt[0])
-        + ", "
-        + str(gt[3])
-        + ", "
-        + str(gt[1])
-        + ", "
-        + str(gt[5] * -1)
-        + ", WGS-84, units=Degrees}"
-    )
-
-    # This creates the coordinate system string
-    # hard-coded replacement of wkt crs could probably be improved, though should be the same for all EMIT datasets
-    csstring = '{ GEOGCS["WGS 84",DATUM["WGS_1984",SPHEROID["WGS 84",6378137,298.257223563,AUTHORITY["EPSG","7030"]],AUTHORITY["EPSG","6326"]],PRIMEM["Greenwich",0,AUTHORITY["EPSG","8901"]],UNIT["degree",0.0174532925199433,AUTHORITY["EPSG","9122"]],AXIS["Latitude",NORTH],AXIS["Longitude",EAST],AUTHORITY["EPSG","4326"]] }'
-    # List data variables (typically reflectance/radiance)
-    var_names = list(xr_ds.data_vars)
-
-    # Loop through variable names
-    for var in var_names:
-        # Define output filename
-        output_name = os.path.join(output_dir, xr_ds.attrs["granule_id"] + "_" + var)
-
-        nbands = 1
-        if len(xr_ds[var].data.shape) > 2:
[…]
