### 2024-08-20 opengeos/HyperCoast#117 @giswqs
> I can't see how this method, in the emit.ipynb notebook is working, or if it works. I tried changing the numbers and nothing changed: m.add("spectral", xlim=(400, 1400))

The `xlim` parameter is used to limit the wavelength range to be displayed on the plot. It works as expected on my end.

## Without xlim
```python
m = hypercoast.Map()
m.add_basemap("SATELLITE")
m.add_emit(dataset, wavelengths=[1000, 600, 500], vmin=0, vmax=0.3, layer_name="EMIT")
m.add("spectral")
m
```
![image](https://github.com/user-attachments/assets/c40be2a2-50f4-4b7a-8e3a-1c430ede5058)

## With xlim
```python
m = hypercoast.Map()
m.add_basemap("SATELLITE")
m.add_emit(dataset, wavelengths=[1000, 600, 500], vmin=0, vmax=0.3, layer_name="EMIT")
m.add("spectral", xlim=(400, 1200))
m
```
![image](https://github.com/user-attachments/assets/7132cb6c-fb47-4de9-8e5b-a970b3992318)
