### 2024-08-20 opengeos/HyperCoast#117 @alexgleith
> wavelength range to be displayed on the plot

Ah, on the plot! I was expecting it to stretch on the map. Now it makes sense that it's `xlim` and not `zlim` for the actual visualisation.
