### 2024-08-20 opengeos/HyperCoast#117 @giswqs
> The function m.spectral_to_df() structures the dataframe in an unexpected way. I think exporting something with a geometry, or with lon/lat columns and 285 more spectral columns would make more sense, even if the structure is a bit clumsy.

The strcture of the dataframe looks like this. It is easier to plot a data strucutre like this with [bqplot](https://bqplot.github.io/bqplot/) (the libary we use for plotting spectral profiles), which takes a dataframe with x and y columns.

However, I do like your suggestion for export something with a geometry. I will create a `spectral_to_gdf` method for exporting a geodataframe. One caveat is that the shapefile format generally does not like >256 columns. Anyway, we can let users decide what vector data format they want to convert the GeoDataFrame to.

![image](https://github.com/user-attachments/assets/e8c1616f-7609-4f59-8c71-76ae3878fa42)
