### 2024-08-20 opengeos/HyperCoast#117 @alexgleith
>The purpose of those three sample files was to make the GitHub Actions unit tests run faster.

This makes sense.

>We do provide a separate notebook example ([search_data.ipynb](https://hypercoast.org/examples/search_data/)) that demonstrates how to download NASA EarthData interactively and programmatically.

It _does_, but it also wraps the earthdata functions in a leafmap function that is used by the separate hypercoast function. I'm not really sure why it does this, and it obfuscates the way that the underlying earthdata API, and Python libraries work.
