### 2024-08-20 opengeos/HyperCoast#117 @alexgleith
> I will create a spectral_to_gdf method for exporting a geodataframe.

Good idea!

> the shapefile format generally does not like >256 columns

Hmm, yeah, a bit of an issue, but it's nice to have an excuse to not support Shapefiles! haha
