### 2024-08-20 opengeos/HyperCoast#117 @giswqs
The `Map.spectral_to_gdf()` method has been added in https://github.com/opengeos/HyperCoast/pull/118
