### 2024-08-20 eed103b
Add spectral_to_gdf method (#118)
