### 2024-08-20 opengeos/HyperCoast#117 @giswqs
> It does, but it also wraps the earthdata functions in a leafmap function that is used by the separate hypercoast function. I'm not really sure why it does this, and it obfuscates the way that the underlying earthdata API, and Python libraries work.

The `leafmap.nasa_data_search()` function wraps the `earthaccess.search_data()` method with additional inprovements, such as returning the data footprints as a GeoDataFrame, or saving it as a vector format (e.g., geojson, shapefile).
https://github.com/opengeos/leafmap/blob/a6dae078146064a870fa2aada5d462259651e3bc/leafmap/common.py#L13296

The reason that hypercoast wraps the leafmap functions is to simplify the import statements in the notebook examples. Most notebook examples only need `import hypercoast` so that users don't need to learn other dependencies.
