### 2024-08-20 #7025 @platipodium
> @platipodium Thanks for paying attention to this. These labels are automatically generated based on the "Software report", since there are Tex files in the repo, the label is added.

Automatic does not mean correct :=). this is not a tex project, so it is misleading.  But we could file a bug with JOSS itself?
