### 2024-08-20 opengeos/HyperCoast#119 @platipodium
Clarify authorship
This is part of the review at https://github.com/openjournals/joss-reviews/issues/7025

- In your paper, you state Wu & Liu as authors.   No-one else is acknowledged
- In your git history, there are small contributions by Leith, Slowy & Ponce-Campos

Please clarify.  You might have a look at [all contributors ](https://allcontributors.org).
