### 2024-08-20 fe9d2a1
Add contributors section to README
