### 2024-08-20 ae8d25d
Add all-contributors bot (#120)
* Add all-contributors bot

* Add contributors section to README
