### 2024-08-20 790e3e3
Add all-contributors bot (#120)
* Add all-contributors bot

* Add contributors section to README
