### 2024-08-20 opengeos/HyperCoast#121 @giswqs
Add contributors to README
Fix #119

https://github.com/openjournals/joss-reviews/issues/7025
