### 2024-08-20 bd49ed7
docs: add bingqing-liu as a contributor for code, design, and ideas (#122)
* docs: update README.md [skip ci]

* docs: create .all-contributorsrc [skip ci]

---------

Co-authored-by: allcontributors[bot] <46447321+allcontributors[bot]@users.noreply.github.com>
