### 2024-08-20 fb65f83
docs: add giswqs as a contributor for code, design, and maintenance (#123)
* docs: update README.md [skip ci]

* docs: update .all-contributorsrc [skip ci]

---------

Co-authored-by: allcontributors[bot] <46447321+allcontributors[bot]@users.noreply.github.com>
