### 2024-08-20 opengeos/HyperCoast#121 @giswqs
@all-contributors please add @alexgleith for code, review
