+++ README.md
@@ -1,6 +1,6 @@
 # HyperCoast
 <!-- ALL-CONTRIBUTORS-BADGE:START - Do not remove or modify this section -->
-[![All Contributors](https://img.shields.io/badge/all_contributors-2-orange.svg?style=flat-square)](#contributors-)
+[![All Contributors](https://img.shields.io/badge/all_contributors-3-orange.svg?style=flat-square)](#contributors-)
 <!-- ALL-CONTRIBUTORS-BADGE:END -->
 
 [![image](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/opengeos/HyperCoast/blob/main)
@@ -106,6 +106,7 @@ HyperCoast is released under the MIT License. However, some of the modules in Hy
     <tr>
       <td align="center" valign="top" width="14.28%"><a href="https://github.com/bingqing-liu"><img src="https://avatars.githubusercontent.com/u/123585527?v=4?s=100" width="100px;" alt="Bingqing Liu"/><br /><sub><b>Bingqing Liu</b></sub></a><br /><a href="https://github.com/opengeos/HyperCoast/commits?author=bingqing-liu" title="Code">💻</a> <a href="#design-bingqing-liu" title="Design">🎨</a> <a href="#ideas-bingqing-liu" title="Ideas, Planning, & Feedback">🤔</a></td>
       <td align="center" valign="top" width="14.28%"><a href="https://gishub.org"><img src="https://avatars.githubusercontent.com/u/5016453?v=4?s=100" width="100px;" alt="Qiusheng Wu"/><br /><sub><b>Qiusheng Wu</b></sub></a><br /><a href="https://github.com/opengeos/HyperCoast/commits?author=giswqs" title="Code">💻</a> <a href="#design-giswqs" title="Design">🎨</a> <a href="#maintenance-giswqs" title="Maintenance">🚧</a></td>
+      <td align="center" valign="top" width="14.28%"><a href="https://auspatious.com"><img src="https://avatars.githubusercontent.com/u/3445853?v=4?s=100" width="100px;" alt="Alex Leith"/><br /><sub><b>Alex Leith</b></sub></a><br /><a href="https://github.com/opengeos/HyperCoast/commits?author=alexgleith" title="Code">💻</a> <a href="https://github.com/opengeos/HyperCoast/pulls?q=is%3Apr+reviewed-by%3Aalexgleith" title="Reviewed Pull Requests">👀</a></td>
     </tr>
   </tbody>
 </table>
