### 2024-08-20 bc57f49
docs: add slowy07 as a contributor for code, and maintenance (#125)
* docs: update README.md [skip ci]

* docs: update .all-contributorsrc [skip ci]

---------

Co-authored-by: allcontributors[bot] <46447321+allcontributors[bot]@users.noreply.github.com>
