### 2024-08-20 d8bdcc7
docs: add gponce-ars as a contributor for code, and bug (#126)
* docs: update README.md [skip ci]

* docs: update .all-contributorsrc [skip ci]

---------

Co-authored-by: allcontributors[bot] <46447321+allcontributors[bot]@users.noreply.github.com>
