### 2024-08-20 opengeos/HyperCoast#119 @giswqs
Thank you for the suggestion. This issue has been addressed in #120 and #121
