### 2024-08-20 opengeos/HyperCoast#121 @giswqs
@all-contributors please add @platipodium for review
