+++ hypercoast/hypercoast.py
@@ -8,13 +8,33 @@ import ipyleaflet
 import leafmap
 import xarray as xr
 import numpy as np
-from .aviris import *
-from .common import *
-from .desis import *
-from .emit import *
-from .neon import *
-from .pace import *
+from typing import Union
+from .aviris import aviris_to_image, read_aviris
+from .desis import desis_to_image, read_desis
+from .emit import emit_to_image, read_emit
+from .neon import neon_to_image, read_neon
+from .pace import pace_to_image, read_pace
 from .ui import SpectralWidget
+from .common import (
+    download_file,
+    search_datasets,
+    search_nasa_data,
+    download_nasa_data,
+    search_pace,
+    search_pace_chla,
+    search_emit,
+    search_ecostress,
+    download_pace,
+    download_emit,
+    download_ecostress,
+    nasa_earth_login,
+    image_cube,
+    open_dataset,
+    download_acolite,
+    run_acolite,
+    pca,
+    show_field_data,
+)
 
 
 class Map(leafmap.Map):
@@ -38,6 +58,9 @@ class Map(leafmap.Map):
                 class's constructor.
         """
         super().__init__(**kwargs)
+        self._spectral_data = {}
+        self._plot_options = None
+        self._plot_marker_cluster = None
 
     def add(self, obj, position="topright", xlim=None, ylim=None, **kwargs):
         """Add a layer to the map.
@@ -107,10 +130,13 @@ class Map(leafmap.Map):
         nodata=None,
         attribution=None,
         layer_name="Raster",
+        layer_index=None,
         zoom_to_layer=True,
         visible=True,
-        array_args={},
-        open_args={},
+        opacity=1.0,
+        array_args=None,
+        client_args={"cors_all": False},
+        open_args=None,
         **kwargs,
     ):
         """Add a local raster dataset to the map.
@@ -142,16 +168,27 @@ class Map(leafmap.Map):
             attribution (str, optional): Attribution for the source raster. This
                 defaults to a message about it being a local file.. Defaults to None.
             layer_name (str, optional): The layer name to use. Defaults to 'Raster'.
+            layer_index (int, optional): The index of the layer. Defaults to None.
             zoom_to_layer (bool, optional): Whether to zoom to the extent of the
                 layer. Defaults to True.
             visible (bool, optional): Whether the layer is visible. Defaults to
                 True.
+            opacity (float, optional): The opacity of the layer. Defaults to 1.0.
             array_args (dict, optional): Additional arguments to pass to
                 `array_to_memory_file` when reading the raster. Defaults to {}.
+            client_args (dict, optional): Additional arguments to pass to
+                localtileserver.TileClient. Defaults to { "cors_all": False }.
+            open_args (dict, optional): Additional arguments to pass to
+                rioxarray.open_rasterio.
+
         """
 
-        import numpy as np
         import rioxarray as rxr
+
+        if array_args is None:
+            array_args = {}
+        if open_args is None:
+            open_args = {}
 
         if nodata is None:
             nodata = np.nan
@@ -164,9 +201,12 @@ class Map(leafmap.Map):
             nodata=nodata,
             attribution=attribution,
             layer_name=layer_name,
+            layer_index=layer_index,
             zoom_to_layer=zoom_to_layer,
             visible=visible,
+            opacity=opacity,
             array_args=array_args,
+            client_args=client_args,
             **kwargs,
         )
 
@@ -191,12 +231,18 @@ class Map(leafmap.Map):
         layer_name="Raster",
         zoom_to_layer=True,
         visible=True,
-        array_args={},
-        open_args={},
[…]
