### 2024-08-20 f6e2195
Add view_pace_pixel_locations import
