### 2024-08-20 7f76844
Fix pylint warnings (#128)
* Fix pylint warnings

* Add view_pace_pixel_locations import

* Fix import errors

* Fix import errors
