### 2024-08-20 opengeos/HyperCoast#117 @giswqs
@alexgleith I have fixed the linter warnings in https://github.com/opengeos/HyperCoast/pull/128. Thank you once again for taking the time to review it and providing constructive feedback.
