+++ hypercoast/__init__.py
@@ -6,7 +6,7 @@
 
 __author__ = """Qiusheng Wu"""
 __email__ = "giswqs@gmail.com"
-__version__ = "0.7.6"
+__version__ = "0.7.7"
 
 
 from .hypercoast import *
