### 2024-08-23 #7025 @cheginit
@giswqs, thank you for diligently addressing the comments raised by the reviewer.

I have some comments on the paper:

- [ ] Line 25: Please use `German Aerospace Center (DLR)`
- [ ] Line 30: Pleas use `earthaccess` instead of earthaccess.
- [ ] Line 59: Please clarify why HyperSpy is not tailored for coastal applications. My understanding is that HyperSpy is somewhat similar to `xarray`. So, I find this sentence vague.
- [ ] Please use a consistent format for paper titles in References. Some references, e.g., the first one, have title case style, and others, like the fourth one, only the first letter is capitalized. Please pick one style and apply to all reference.
- [ ] References on lines 93 and 134 are missing journal name.
