### 2024-08-23 f8e9079
Fix JOSS paper review comments
