+++ paper/paper.bib
@@ -1,7 +1,5 @@
-
 @ARTICLE{Wu2021,
-  title     = "{Leafmap: A Python package for interactive mapping and geospatial
-               analysis with minimal coding in a Jupyter environment}",
+  title     = "{Leafmap: A Python Package for Interactive Mapping and Geospatial Analysis with Minimal Coding in a Jupyter Environment}",
   author    = "Wu, Qiusheng",
   journal   = "Journal of Open Source Software",
   publisher = "joss.theoj.org",
@@ -11,10 +9,9 @@
 }
 
 @ARTICLE{Sullivan2019,
-  title     = "{PyVista: 3D plotting and mesh analysis through a streamlined
-               interface for the Visualization Toolkit (VTK)}",
+  title     = "{PyVista: 3D Plotting and Mesh Analysis Through a Streamlined Interface for the Visualization Toolkit (VTK)}",
   author    = "Sullivan, C and Kaszynski, Alexander",
-  journal   = "Journal of open source software",
+  journal   = "Journal of Open Source Software",
   publisher = "The Open Journal",
   volume    =  4,
   number    =  37,
@@ -27,17 +24,8 @@
 }
 
 @MISC{De_La_Pena2017,
-  title     = "{hyperspy/hyperspy: HyperSpy 1.3}",
-  author    = "De La Pe{\~n}a, Francisco and Ostasevicius, Tomas and Tonaas
-               Fauske, Vidar and Burdet, Pierre and Jokubauskas, Petras and
-               Nord, Magnus and Prestat, Eric and Sarahan, Mike and MacArthur,
-               Katherine E and Johnstone, Duncan N and Taillon, Joshua and
-               Caron, Jan and Furnival, Tom and Eljarrat, Alberto and Mazzucco,
-               Stefano and Migunov, Vadim and Aarholt, Thomas and Walls,
-               Michael and Winkler, Florian and Martineau, Ben and Donval,
-               Ga{\"e}l and Hoglund, Eric R and Alxneit, Ivo and Hjorth, Ida
-               and Zagonel, Luiz Fernando and Garmannslund, Andreas and Gohlke,
-               Christoph and Iyengar, Ilya and Chang, Huang-Wei",
+  title     = "{Hyperspy/Hyperspy: HyperSpy 1.3}",
+  author    = "De La Pe{\~n}a, Francisco and Ostasevicius, Tomas and Tonaas Fauske, Vidar and Burdet, Pierre and Jokubauskas, Petras and Nord, Magnus and Prestat, Eric and Sarahan, Mike and MacArthur, Katherine E and Johnstone, Duncan N and Taillon, Joshua and Caron, Jan and Furnival, Tom and Eljarrat, Alberto and Mazzucco, Stefano and Migunov, Vadim and Aarholt, Thomas and Walls, Michael and Winkler, Florian and Martineau, Ben and Donval, Ga{\"e}l and Hoglund, Eric R and Alxneit, Ivo and Hjorth, Ida and Zagonel, Luiz Fernando and Garmannslund, Andreas and Gohlke, Christoph and Iyengar, Ilya and Chang, Huang-Wei",
   journal   = "Zenodo",
   publisher = "Zenodo",
   year      =  2017,
@@ -46,31 +34,23 @@
 }
 
 @INPROCEEDINGS{Gorman2019,
-  title      = "{The NASA Plankton, Aerosol, Cloud, ocean Ecosystem (PACE)
-                mission: an emerging era of global, hyperspectral Earth system
-                remote sensing}",
+  title      = "{The NASA Plankton, Aerosol, Cloud, Ocean Ecosystem (PACE) Mission: An Emerging Era of Global, Hyperspectral Earth System Remote Sensing}",
   booktitle  = "{Sensors, Systems, and Next-Generation Satellites XXIII}",
[…]
