+++ paper/paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 
 HyperCoast is a Python package designed to provide an accessible and comprehensive set of tools for visualizing and analyzing hyperspectral data in coastal environments. Hyperspectral data refers to the information collected by sensors that capture light across a wide range of wavelengths, beyond what the human eye can see. This data allows scientists to detect and analyze various materials and conditions on the Earth's surface with great detail. Unlike multispectral data, which captures light in a limited number of broad wavelength bands (typically 3 to 10), hyperspectral data captures light in many narrow, contiguous wavelength bands, often numbering in the hundreds. This provides much more detailed spectral information. Leveraging the capabilities of popular packages like Leafmap [@Wu2021] and PyVista [@Sullivan2019], HyperCoast streamlines the exploration and interpretation of complex hyperspectral remote sensing data from existing spaceborne and airborne missions. It is also poised to support future hyperspectral missions, such as NASA's SBG and GLIMR [@Dierssen2021].
 
-HyperCoast supports the reading and visualization of hyperspectral data from various missions, including the Airborne Visible/Infrared Imaging Spectrometer (AVIRIS) [@Green1998], the National Ecological Observatory Network (NEON) Airborne Observation Platform (AOP) [@Kampe2010], the Plankton, Aerosol, Cloud, ocean Ecosystem (PACE) mission [@Gorman2019], the Earth Surface Mineral Dust Source Investigation (EMIT) [@Green2021], and the DLR Earth Sensing Imaging Spectrometer (DESIS) [@Alonso2019], along with other datasets like the ECOsystem Spaceborne Thermal Radiometer Experiment on Space Station (ECOSTRESS) [@Fisher2020]. Users can interactively explore hyperspectral data, extract spectral signatures, change band combinations and colormaps, visualize data in 3D, and perform interactive slicing and thresholding operations (see Figure 1). Additionally, by leveraging the earthaccess [@barrett2024] package, HyperCoast provides tools for interactively searching NASA's hyperspectral data. This makes HyperCoast a versatile and powerful tool for working with hyperspectral data globally, with a particular focus on coastal regions.
+HyperCoast supports the reading and visualization of hyperspectral data from various missions, including the Airborne Visible/Infrared Imaging Spectrometer (AVIRIS) [@Green1998], the National Ecological Observatory Network (NEON) Airborne Observation Platform (AOP) [@Kampe2010], the Plankton, Aerosol, Cloud, ocean Ecosystem (PACE) mission [@Gorman2019], the Earth Surface Mineral Dust Source Investigation (EMIT) [@Green2021], and the German Aerospace Center (DLR) Earth Sensing Imaging Spectrometer (DESIS) [@Alonso2019], along with other datasets like the ECOsystem Spaceborne Thermal Radiometer Experiment on Space Station (ECOSTRESS) [@Fisher2020]. Users can interactively explore hyperspectral data, extract spectral signatures, change band combinations and colormaps, visualize data in 3D, and perform interactive slicing and thresholding operations (see Figure 1). Additionally, by leveraging the `earthaccess` [@barrett2024] package, HyperCoast provides tools for interactively searching NASA's hyperspectral data. This makes HyperCoast a versatile and powerful tool for working with hyperspectral data globally, with a particular focus on coastal regions.
 
 ![EMIT](https://assets.gishub.org/images/EMIT-demo.png)
 **Figure 1.** An example of visualizing NASA EMIT hyperspectral data using HyperCoast.
@@ -39,7 +39,7 @@ Coastal systems, characterized by complex physical, chemical, and bio-optical pr
 
[…]
