### 2024-08-23 0004956
Bump version: 0.7.7 → 0.7.8
