### 2024-08-23 #7025 @giswqs
@cheginit Thank you for the feedback. All your comments have been addressed in https://github.com/opengeos/HyperCoast/pull/129.

Latest release: v0.7.8
Zenodo DOI: https://doi.org/10.5281/zenodo.13368024
