### 2024-08-26 #7025 @cheginit
@giswqs Thanks for quickly addressing the comments, we're almost there! Please use the same title for the Zenodo archive as of this submission, i.e., "HyperCoast: A Python Package for Visualizing and Analyzing Hyperspectral Data in Coastal Environments".
