### 2024-08-26 #7025 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
