### 2024-08-26 2fb7031
Add JOSS paper
