+++ docs/index.md
@@ -21,6 +21,12 @@ HyperCoast supports the reading and visualization of hyperspectral data from var
 ![EMIT](https://assets.gishub.org/images/EMIT-demo.png)
 **Figure 1.** An example of visualizing NASA EMIT hyperspectral data using HyperCoast.
 
+## Citations
+
+If you find HyperCoast useful in your research, please consider citing the following papers to support us. Thank you!
+
+-   Liu, B. & Wu, Q. (2024). HyperCoast: A Python Package for Visualizing and Analyzing Hyperspectral Data in Coastal Environments. _Journal of Open Source Software_, 9(100), 7025. <https://doi.org/10.21105/joss.07025>.
+
 ## Features
 
 -   Searching for NASA hyperspectral data interactively
