### 2024-08-26 cc02fa9
Add JOSS paper to docs
