### 2024-08-26 851f8ef
Add JOSS paper (#131)
* Add JOSS paper

* Add JOSS paper to docs
