### 2024-08-26 3e5a307
Add CITATION.cff (#132)
