+++ CITATION.cff
@@ -1,0 +1,34 @@
+cff-version: "1.2.0"
+authors:
+    - family-names: Liu
+      given-names: Bingqing
+      orcid: "https://orcid.org/0000-0003-4651-6996"
+    - family-names: Wu
+      given-names: Qiusheng
+      orcid: "https://orcid.org/0000-0001-5437-4073"
+doi: 10.5281/zenodo.13368024
+message: If you use this software, please cite our article in the
+    Journal of Open Source Software.
+preferred-citation:
+    authors:
+        - family-names: Liu
+          given-names: Bingqing
+          orcid: "https://orcid.org/0000-0003-4651-6996"
+        - family-names: Wu
+          given-names: Qiusheng
+          orcid: "https://orcid.org/0000-0001-5437-4073"
+    date-published: 2024-08-26
+    doi: 10.21105/joss.07025
+    issn: 2475-9066
+    issue: 100
+    journal: Journal of Open Source Software
+    publisher:
+        name: Open Journals
+    start: 7025
+    title: "HyperCoast: A Python Package for Visualizing and Analyzing
+        Hyperspectral Data in Coastal Environments"
+    type: article
+    url: "https://joss.theoj.org/papers/10.21105/joss.07025"
+    volume: 9
+title: "HyperCoast: A Python Package for Visualizing and Analyzing
+    Hyperspectral Data in Coastal Environments"
